+++ tests/subtract_and_print.c
@@ -2,7 +2,7 @@
 
 // TEST:{ "func": "main", "output": "Z" }
 int main() {
-  int a = 'z';
-  int b = 0x20;
+  uint8_t a = 'z';
+  uint8_t b = 0x20;
   putc(a - b);
 };
